+++ libnail/src/align/traceback.rs
@@ -1,4 +1,5 @@
 use crate::align::structs::{DpMatrix, Trace};
+use crate::structs::profile::Transition;
 use crate::structs::Profile;
 
 pub fn traceback(
@@ -22,15 +23,16 @@ pub fn traceback(
     while previous_state != Trace::S_STATE {
         current_state = match previous_state {
             Trace::C_STATE => {
-                let c_to_c_path = profile
-                    .special_transition_score_delta(Profile::C_IDX, Profile::SPECIAL_LOOP_IDX)
-                    * (optimal_matrix
+                let c_to_c_path = profile.special_transition_score_delta(
+                    Profile::C_IDX,
+                    Profile::SPECIAL_LOOP_IDX as usize,
+                ) * (optimal_matrix
                     .get_special(target_idx - 1, Profile::C_IDX)
                     // TODO: why does this specific path involve a posterior probability?
                     + posterior_matrix.get_special(target_idx, Profile::C_IDX));
 
                 let c_to_e_path = profile
-                    .transition_score_delta(Profile::E_IDX, Profile::SPECIAL_MOVE_IDX)
+                    .transition_score_delta(Profile::E_IDX, Profile::SPECIAL_MOVE_IDX as usize)
                     * optimal_matrix.get_special(target_idx, Profile::E_IDX);
 
                 if c_to_c_path > c_to_e_path {
@@ -70,13 +72,13 @@ pub fn traceback(
                 ];
 
                 let possible_paths: [f32; 4] = [
-                    profile.transition_score_delta(Profile::M_M_IDX, profile_idx - 1)
+                    profile.transition_score_delta(Transition::MM as usize, profile_idx - 1)
                         * optimal_matrix.get_match(target_idx - 1, profile_idx - 1),
-                    profile.transition_score_delta(Profile::I_M_IDX, profile_idx - 1)
+                    profile.transition_score_delta(Transition::IM as usize, profile_idx - 1)
                         * optimal_matrix.get_insert(target_idx - 1, profile_idx - 1),
-                    profile.transition_score_delta(Profile::D_M_IDX, profile_idx - 1)
+                    profile.transition_score_delta(Transition::DM as usize, profile_idx - 1)
                         * optimal_matrix.get_delete(target_idx - 1, profile_idx - 1),
-                    profile.transition_score_delta(Profile::B_M_IDX, profile_idx - 1)
+                    profile.transition_score_delta(Transition::BM as usize, profile_idx - 1)
                         * optimal_matrix.get_special(target_idx - 1, Profile::B_IDX),
                 ];
 
@@ -95,11 +97,11 @@ pub fn traceback(
             }
             Trace::I_STATE => {
                 let match_to_insert_path = profile
-                    .transition_score_delta(Profile::M_I_IDX, profile_idx)
+                    .transition_score_delta(Transition::MI as usize, profile_idx)
                     * optimal_matrix.get_match(target_idx - 1, profile_idx);
 
                 let insert_to_insert_path: f32 = profile
-                    .transition_score_delta(Profile::I_I_IDX, profile_idx)
+                    .transition_score_delta(Transition::II as usize, profile_idx)
                     * optimal_matrix.get_insert(target_idx - 1, profile_idx);
 
                 // an insert means we moved forward only in the profile
@@ -113,11 +115,11 @@ pub fn traceback(
             }
             Trace::D_STATE => {
                 let match_to_delete_path = profile
-                    .transition_score_delta(Profile::M_D_IDX, profile_idx - 1)
+                    .transition_score_delta(Transition::MD as usize, profile_idx - 1)
                     * optimal_matrix.get_match(target_idx, profile_idx - 1);
 
                 let delete_to_delete_path = profile
-                    .transition_score_delta(Profile::D_D_IDX, profile_idx - 1)
+                    .transition_score_delta(Transition::DD as usize, profile_idx - 1)
                     * optimal_matrix.get_delete(target_idx, profile_idx - 1);
 
                 // a delete means we moved forward only in the profile
@@ -130,13 +132,15 @@ pub fn traceback(
                 }
             }
             Trace::B_STATE => {
-                let n_to_b_path = profile
-                    .special_transition_score_delta(Profile::N_IDX, Profile::SPECIAL_MOVE_IDX)
-                    * optimal_matrix.get_special(target_idx, Profile::N_IDX);
+                let n_to_b_path = profile.special_transition_score_delta(
+                    Profile::N_IDX,
+                    Profile::SPECIAL_MOVE_IDX as usize,
+                ) * optimal_matrix.get_special(target_idx, Profile::N_IDX);
 
-                let j_to_b_path = profile
-                    .special_transition_score_delta(Profile::J_IDX, Profile::SPECIAL_MOVE_IDX)
-                    * optimal_matrix.get_special(target_idx, Profile::J_IDX);
+                let j_to_b_path = profile.special_transition_score_delta(
+                    Profile::J_IDX,
+                    Profile::SPECIAL_MOVE_IDX as usize,
+                ) * optimal_matrix.get_special(target_idx, Profile::J_IDX);
 
                 if n_to_b_path >= j_to_b_path {
                     Trace::N_STATE
@@ -152,16 +156,18 @@ pub fn traceback(
                 }
             }
             Trace::J_STATE => {
-                let j_to_j_path = profile
-                    .special_transition_score_delta(Profile::J_IDX, Profile::SPECIAL_LOOP_IDX)
-                    * (optimal_matrix
+                let j_to_j_path = profile.special_transition_score_delta(
+                    Profile::J_IDX,
+                    Profile::SPECIAL_LOOP_IDX as usize,
+                ) * (optimal_matrix
                     .get_special(target_idx - 1, Profile::J_IDX)
                     // TODO: why does this specific path involve a posterior probability?
                     + posterior_matrix.get_special(target_idx, Profile::J_IDX));
 
-                let e_to_j_path = profile
-                    .special_transition_score_delta(Profile::E_IDX, Profile::SPECIAL_LOOP_IDX)
-                    * optimal_matrix.get_special(target_idx, Profile::E_IDX);
+                let e_to_j_path = profile.special_transition_score_delta(
+                    Profile::E_IDX,
+                    Profile::SPECIAL_LOOP_IDX as usize,
+                ) * optimal_matrix.get_special(target_idx, Profile::E_IDX);
 
                 if j_to_j_path > e_to_j_path {
                     Trace::J_STATE