+++ nail/src/io/mod.rs
@@ -5,4 +5,5 @@ mod p7hmm;
 pub use p7hmm::*;
 
 mod impl_rayon;
+
 mod util;