+++ libnail/CHANGELOG.md
@@ -17,14 +17,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- added `AmbiguityMap` struct
-- added `AminoUtilsUtf8` and `AminoUtilsDigital` traits
+- added struct `AmbiguityMap`
+- added traits `AminoUtilsUtf8` and `AminoUtilsDigital`
+- added enum `Transition` 
+- added mod `profile::blosum62`
+- added struct `ProfileBuilder`
+- added function `Profile::from_blosum62_and_seq()`
 
 ### Changed
+- removed transition index constants under `Profile` namespace
 - renamed `Profile.forward_tau` and `Profile.forward_lambda` to `fwd_tau` and `fwd_lambda`
 - renamed `Profile.consensus_sequence_bytes_utf8` to `consensus_seq_bytes_utf8`
 - moved `Alphabet` and`AminoAcid` to mod `alphabet`
 - split self mutating functions from trait `VecMath` into new trait `VecMathMut`
+
+### Removed
+- removed `Hmm` struct
+- removed function `Profile::new()`
 
 ## [0.4.0] - 2025-6-18
 
@@ -66,7 +75,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - the mean relative entropy (MRE) can now be raised or lowered
     - the algorithm for raising (MRE) has been improved
 - `util::avg_relative_entropy()` is now `mean_relative_entropy()`
-- parse_hmms_from_p7hmm_file(path) is now Hmm::from_p7hmm(buf)
+- `parse_hmms_from_p7hmm_file(path)` is now `Hmm::from_p7hmm(buf)`
 - refactored Alignment struct
     - fields are now grouped into sub-structs
     - refactored AlignmentBuilder for changes